+++ src/index.ts
@@ -7,9 +7,14 @@
 export { countries } from './data/countries.js';
 
 export type { Country, Currency, NumberFormat } from './types/country.js';
-export type { LocaleOption, NumberFormatOption } from './types/option.js';
+export type { DateFormatOption, LocaleOption, NumberFormatOption } from './types/option.js';
 
-export { countryOptions, currencyOptions, numberFormatOptions } from './options.js';
+export {
+  countryOptions,
+  currencyOptions,
+  dateFormatOptions,
+  numberFormatOptions,
+} from './options.js';
 
 export { getCountryInfo } from './utils/getCountryInfo.js';
 export { formatCurrency } from './utils/formatCurrency.js';